Fixes #1 by handling dig's output better.
A bug prevented this cmdlet from calling `dig` successfully when a
user-specified -Server was included. This was resolved by changing
how `dig` is called, so that whitespace is parsed correctly by the
PowerShell runtime.  Thanks, GitCouth!

This commit also moves the calls to `dig` into a separate function,
for portability and easier maintenance in the future. Debug output
is also available via PowerShell's debug stream.

In addition, `dig` can be found and executed even if the binary is
not in its default location of `/usr/bin`.  However, it still must
be located in the shell's PATH.